feat: add initial source code
A lot of tests are missing and will gradually be added later on. The current functionality is only tested locally so there might still be some bugs in here. We just have to live with that at the moment, and of course issues should be created if a bug is found.