feat!: format time using Luxon instead of Moment
Replaces the `time.local` option with `time.utc`, which is essentially inverted. The `time.type` option is replaced with the `time.preset` option. This also introduces the Luxon formatting presets, which gives more preset formats. As a consequence, the `time.format` option syntax is slightly different then it was with Moment.